いいねとお気に入り機能のmember_idが取得できないよエラーとおさらば これ、革命です

+++ app/Livewire/Favorite.php
@@ -3,8 +3,6 @@
 namespace App\Livewire;
 use App\Models\Favorite as FavoriteModel;
 use Livewire\Component;
-use Auth;
-
 class Favorite extends Component
 {
     public $file, $storeID;
@@ -38,33 +36,42 @@ class Favorite extends Component
         return view('livewire.favorite');
 
     }
-    public function toggleFavorite(){
 
+    public function toggleFavorite(){
         // Favoriteモデルからmember_idとstore_idが一致するデータを取得　... ①
         $memberID = session()->get('member_id'); 
-        $favorite = FavoriteModel::withTrashed()->where('member_id', $memberID)
-                  ->where('store_id', $this->storeID)->first();
 
-        // ①のデータが存在しない場合
-            // 新規作成（INSERT）
-            if ($favorite === null) {
-                FavoriteModel::create([
-                    'member_id' => $memberID,
-                    'store_id' => $this->storeID,
-                ]);
-                
-        // ①が存在する&&deleted_atがnullの場合 ... ②
-            // 現在日時をセットする（UPDATE）→ 削除扱い
-        } elseif (!$favorite->trashed()) {
-            $favorite->delete();
+    // ログインしているときにお気に入りしたとき
+        if ($memberID !== null) {
+            $favorite = FavoriteModel::withTrashed()->where('member_id', $memberID)
+                    ->where('store_id', $this->storeID)->first();
 
-        // ②以外の場合
-            // deletedt_atをnullにする（UPDATE）→ 再度ブックマークした扱い
+            // ①のデータが存在しない場合
+                // 新規作成（INSERT）
+                if ($favorite === null) {
+                    FavoriteModel::create([
+                        'member_id' => $memberID,
+                        'store_id' => $this->storeID,
+                    ]);
+                    
+            // ①が存在する&&deleted_atがnullの場合 ... ②
+                // 現在日時をセットする（UPDATE）→ 削除扱い
+            } elseif (!$favorite->trashed()) {
+                $favorite->delete();
+
+            // ②以外の場合
+                // deletedt_atをnullにする（UPDATE）→ 再度ブックマークした扱い
+            } else {
+                $favorite->restore();
+            }
+
+            $this->file = $this->isFavorite() ? 'bookmark02@2x.png' : 'bookmark01@2x.png';
+
+    // ログインしていない時にお気に入りしたときのエラー回避
         } else {
-            $favorite->restore();
+            session()->flash('message', 'お気に入りに登録する場合はログインしてください。');
+            return redirect()->route('user.detail-main', ['id' => $this->storeID]);
         }
-
-        $this->file = $this->isFavorite() ? 'bookmark02@2x.png' : 'bookmark01@2x.png';
     }
 
     private function isFavorite(): bool
